+++ NEWS.md
@@ -4,7 +4,8 @@
 - `transmute` is no longer an alias for `select`. This is to make the latter more flexible in some
   bare selection cases (like combining several `tidyselect` calls), and leave the former more simple
   and preferable for actual transmutation cases.
-- `filter_on` gains a `.negate` parameter.
+- `filter_on` gained a `.negate` parameter.
+- `mutate` gained a `.sequential` parameter to enable usage of newly created columns.
 
 # table.express 0.1.1